jjplot-package {jjplot}	R Documentation

The jjplot plotting library for making beautiful plots fast and simple.

Description

jjplot is a plotting library inspired by ggplot2. jjplot is written in grid and while not yet as functional as ggplot2, jjplot can quickly and easily produce beautiful graphics including line graphs, scatter plots, bar graphs, and boxplots. jjplot also makes it straightfoward to compute and overlay statistics and fits onto graphs, and to facet graphs by factor via subplots and automatic color scales.

Details

Package: 	 jjplot

Type: 	 Package

Version: 	 1.0

Date: 	 2010-02-24

License: 	 LGPL

LazyLoad: 	 yes

Full up-to-date documentation with examples can be found at: http://code.google.com/p/jjplot/wiki/Documentation

Author(s)

Jonathan Chang <jcone@princeton.edu> Eytan Bakshy <ebakshy@gmail.com>

References

The ggplot2 package is the inpsiration for (and a more feature-rich version of) this package.

See Also

jjplot

Examples

 ## See the demos for example usages.
 ## Not run: demo(jjplot)

 ## See the following demo for speed tests.
 ## Not run: demo(jjplot.performance)

[Package jjplot version 0.2 Index]